# packages/domain/generated/openapi.yaml
...
openapi: 3.1.0
info:
  title: domain.prim.sh API
  version: 1.0.0
  description: Register domains, manage DNS, auto-TLS. Full domain lifecycle via API.
servers:
  - url: https://domain.prim.sh
security:
  - x402: []
components:
  securitySchemes:
    x402:
      type: apiKey
      in: header
      name: Payment-Signature
      description: |
        EIP-3009 signed payment authorization. Server issues a 402 challenge; client signs and retries.
        See https://prim.sh/pay for the full x402 payment flow.
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - not_found
                - forbidden
                - invalid_request
                - cloudflare_error
                - rate_limited
                - domain_taken
                - quote_expired
                - registrar_error
                - registration_failed
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error message.
...
    CreateZoneRequest:
      type: object
      properties:
        domain:
          type: string
          description: Domain name to create a zone for (e.g. "example.com").
      required:
        - domain
    CreateZoneResponse:
      type: object
      properties:
        zone:
          description: The created zone.
          $ref: "#/components/schemas/GetZoneResponse"
      required:
        - zone
...
    QuoteRequest:
      type: object
      properties:
        domain:
          type: string
          description: Domain name to quote (e.g. "example.com").
        years:
          description: Number of years to register. Default 1.
          type: number
      required:
        - domain
...
    SearchDomainResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/DomainSearchResult"
          description: Search results for each queried domain.
      required:
        - results
...
paths:
...
  /v1/domains/search:
    get:
      operationId: searchDomain
      summary: Check availability and pricing for a domain query
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: query
          in: query
          description: Domain name or keyword to search
          schema:
            type: string
        - name: tlds
          in: query
          description: Comma-separated TLDs (e.g. com,xyz,io)
          schema:
            type: string
      responses:
        "200":
          description: Check availability and pricing for a domain query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchDomainResponse"
        "402":
          description: x402 payment required.
        "502":
          description: NameSilo API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/domains/quote:
    post:
      operationId: quote
      summary: Get a 15-minute price quote for a domain
      security:
        - x402: []
      x-price: $0.001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequest"
      responses:
        "200":
          description: Get a 15-minute price quote for a domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400":
          description: Missing domain or invalid years
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: x402 payment required.
        "502":
          description: NameSilo API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/domains/{domain}/status:
    get:
      operationId: getDomainStatus
      summary: Full post-registration pipeline status (ns_propagated, zone_active, all_ready)
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: domain
          in: path
          required: true
          description: domain parameter
          schema:
            type: string
      responses:
        "200":
          description: Full post-registration pipeline status (ns_propagated, zone_active, all_ready)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRegistrationStatusResponse"
        "402":
          description: x402 payment required.
        "404":
          description: Domain not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones:
    post:
      operationId: createZone
      summary: Create a Cloudflare DNS zone. Returns nameservers to set at your registrar.
      security:
        - x402: []
      x-price: $0.001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneRequest"
      responses:
        "201":
          description: Create a Cloudflare DNS zone. Returns nameservers to set at your registrar.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateZoneResponse"
        "400":
          description: Missing domain or invalid domain name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: x402 payment required.
        "409":
          description: Zone already exists for this domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Cloudflare API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listZones
      summary: List DNS zones owned by the calling wallet (paginated)
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: limit
          in: query
          description: 1-100, default 20
          schema:
            type: integer
        - name: after
          in: query
          description: Cursor from previous response
          schema:
            type: string
      responses:
        "200":
          description: List DNS zones owned by the calling wallet (paginated)
          content:
            application/json:
              schema:
                type: object
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones/{id}:
    get:
      operationId: getZone
      summary: Get zone details
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: id
          in: path
          required: true
          description: id parameter
          schema:
            type: string
      responses:
        "200":
          description: Get zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetZoneResponse"
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteZone
      summary: Delete zone and all records. Irreversible.
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: id
          in: path
          required: true
          description: id parameter
          schema:
            type: string
      responses:
        "200":
          description: Delete zone and all records. Irreversible.
          content:
            application/json:
              schema:
                type: object
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Cloudflare API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones/{zone_id}/activate:
    put:
      operationId: activateZone
      summary: Request Cloudflare NS re-check for faster activation
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: zone_id
          in: path
          required: true
          description: zone_id parameter
          schema:
            type: string
      responses:
        "200":
          description: Request Cloudflare NS re-check for faster activation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivateDomainResponse"
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many zone_activate calls
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones/{zone_id}/verify:
    get:
      operationId: verifyZone
      summary: Check DNS propagation for all zone records
      security:
        - x402: []
      x-price: $0.001
      parameters:
        - name: zone_id
          in: path
          required: true
          description: zone_id parameter
          schema:
            type: string
      responses:
        "200":
          description: Check DNS propagation for all zone records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyDomainResponse"
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones/{zone_id}/mail-setup:
    post:
      operationId: setupMail
      summary: Configure MX, SPF, DMARC, DKIM in one call. Idempotent.
      security:
        - x402: []
      x-price: $0.005
      parameters:
        - name: zone_id
          in: path
          required: true
          description: zone_id parameter
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupMailRequest"
      responses:
        "200":
          description: Configure MX, SPF, DMARC, DKIM in one call. Idempotent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupMailResponse"
        "400":
          description: Missing required mail server fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Cloudflare API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/zones/{zone_id}/records/batch:
    post:
      operationId: batchRecords
      summary: Create, update, and delete DNS records in one atomic request
      security:
        - x402: []
      x-price: $0.005
      parameters:
        - name: zone_id
          in: path
          required: true
          description: zone_id parameter
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRecordsRequest"
      responses:
        "200":
          description: Create, update, and delete DNS records in one atomic request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchRecordsResponse"
        "400":
          description: Invalid record fields or values
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "402":
          description: x402 payment required.
        "403":
          description: Resource belongs to a different wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Zone or record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Cloudflare API error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...